+++ identity-wallet/setler-cli/src/actions/auth.js
@@ -626,7 +626,6 @@ const exec = async (context) => {
       }
 
       log("");
-
       // send the transaction
       const sendPromise = context.coins.sendDirectoryPayment({
         network,