+++ identity-wallet/setler-cli/src/actions/wallet.js
@@ -4,6 +4,10 @@ import prompts from "prompts";
 import qrcode from "qrcode-terminal";
 import { stringToColorBlocks } from "../lib/colorize.js";
 import { waitFor } from "../lib/wait.js";
+import { xrpToDrops } from "xrpl";
+import { getExchangeRate } from "../lib/wallet/getExchangeRate.js";
+import windowSize from "window-size";
+import sysOpen from "open";
 
 const log = console.log;
 
@@ -50,6 +54,245 @@ const exec = async (context) => {
         log(chalk.green(`Wallet initialized`));
       }
       break;
+    case "send": {
+      await gatekeep(context);
+
+      let network =
+        context.flags.network || context.config.network || "xrpl:testnet";
+      if (network === "testnet") {
+        network = "xrpl:testnet";
+      }
+      const keys = await context.vault.keys();
+      let sourceAddress;
+      const networkParts = network.split(":");
+      if (networkParts.length === 1) {
+        sourceAddress = keys[network].address;
+      } else {
+        sourceAddress = keys[networkParts[0]][networkParts[1]].address;
+      }
+
+      // show the user the address we're sending from
+      log(
+        chalk.bold(
+          `\nSending from: ` +
+            chalk.blue(
+              `${sourceAddress}\n` +
+                " ".repeat("sending from: ".length) +
+                stringToColorBlocks(sourceAddress, network)
+            )
+        )
+      );
+      log("");
+
+      // input[2] could be {address}, in which case we send to that address
+      // ask for the address
+      let destinationAddress = context.flags.destination || context.input[2];
+      if (!destinationAddress) {
+        const response = await prompts([
+          {
+            type: "text",
+            name: "address",
+            message: `Enter the address you want to send to: `,
+            initial: false,
+          },
+        ]);
+        if (!response.address) {
+          process.exit(1);
+        }
+        destinationAddress = response.address;
+      }
+
+      // show the user the address we're sending from
+      log(
+        chalk.bold(
+          `\nSending to: ` +
+            chalk.yellow(
+              `${destinationAddress}\n` +
+                " ".repeat("sending to: ".length) +
+                stringToColorBlocks(destinationAddress, network)
+            )
+        )
+      );
+      log("");
+
+      // ask for the amount
+      let amount = context.flags.amount || context.input[3];
+      if (!amount) {
+        const response = await prompts([
+          {
+            type: "text",
+            name: "amount",
+            message: `Enter the amount you want to send (in XRP): `,
+            initial: amount || "0.0001",
+          },
+        ]);
+        if (!response.amount) {
+          process.exit(1);
+        }
+        amount = response.amount;
+      }
+      const amountXrp = amount;
+      const drops = xrpToDrops(amountXrp);
+
+      // ask for the tag
+      let tag = context.flags.tag;
+      if (!tag) {
+        const response = await prompts([
+          {
+            type: "text",
+            name: "tag",
+            message: `Enter the destination tag for this transaction, or leave blank: `,
+            initial: false,
+          },
+        ]);
+        tag = response.tag;
+      }
+
+      // estimate the fees
+      const estPromise = context.coins.estimatedSendFee({
+        network,
+        sourceAddress, // estimate
+        address: sourceAddress, // estimate
+        amount: amountXrp,
+        amountDrops: drops,
+      });
+
+      const estimatedFees = await waitFor(estPromise, {
+        text: `Estimating fees`,
+      });
+      log(
+        chalk.gray(`\tEstimated fees : \t${(estimatedFees * 1).toFixed(6)} XRP`)
+      );
+
+      // convert the amount into usd
+      const getExchange = getExchangeRate("XRP");
+      const exchangeRate = await waitFor(getExchange, {
+        text: "Fetching current exchange rate",
+      });
+
+      const amountUsd = parseFloat(amountXrp * exchangeRate).toFixed(2);
+      log(chalk.gray(`\tAmount in usd  : \t$${amountUsd}\n`));
+
+      // confirm
+      const confirm = await prompts([
+        {
+          type: "confirm",
+          name: "ok",
+          message: `Confirm amount to send: ${amountXrp} XRP?`,
+          initial: false,
+        },
+      ]);
+      if (!confirm.ok) {
+        process.exit(1);
+      }
+      const getAcctPromise = context.coins.getAccountInfo({
+        network,
+        sourceAddress,
+      });
+
+      const accountInfo = await waitFor(getAcctPromise, {
+        text:
+          `Getting account balance for: ` +
+          chalk.yellow(
+            `${sourceAddress}\n` +
+              "                               " +
+              stringToColorBlocks(sourceAddress, network)
+          ),
+      });
+      const balance = accountInfo?.xrpDrops;
+      if (!balance) {
+        log(chalk.red(`send: Could not get balance for ${sourceAddress}`), {
+          balance,
+          accountInfo,
+        });
+        process.exit(1);
+      }
+      const balanceXrp = parseFloat(balance) / 1000000;
+      if (balanceXrp < amountXrp) {
+        log(
+          chalk.red(
+            `send: Not enough funds in ${sourceAddress} to send ${amountXrp} XRP`
+          )
+        );
+        process.exit(1);
+      }
+      log("");
+      log(
+        `Balance after sending ` +
+          chalk.green(`${amountXrp} XRP`) +
+          ` will be ` +
+          chalk.green(`${balanceXrp - amountXrp} XRP`)
+      );
+      log("");
+
+      log("");
+      const { width } = windowSize.get();
+      log(" " + "─".repeat(width - 2));
+      log("");
+
+      const confirm2 = await prompts([
+        {
+          type: "text",
+          name: "doit",
+          message: `Type: ` + chalk.red("send it") + ` to initiate transfer: `,
+          initial: false,
+        },
+      ]);
+      if (confirm2.doit !== "send it") {
+        log(chalk.red(`send: aborting`));
+        process.exit(1);
+      }
+
+      const sendPromise = context.coins.send({
+        network,
+        sourceAddress,
+        address: destinationAddress,
+        amount: amountXrp,
+        amountDrops: drops,
+        tag,
+      });
+      const status = await waitFor(sendPromise, {
+        text:
+          `Sending ` +
+          chalk.green(`${amount} XRP`) +
+          ` to ` +
+          chalk.blue(`${destinationAddress}`) +
+          ` on ` +
+          chalk.magenta(`${network}\n`) +
+          " ".repeat(
+            (`Sending ` + chalk.green(`${amount} XRP`) + ` to `).length
+          ) +
+          "\n",
+      });
+
+      if (status) {
+        if (context.debug) {
+          log(chalk.magenta(JSON.stringify(status, null, "  ")));
+        }
+        log(chalk.bold(`Ok, sent.`));
+      } else {
+        log(chalk.red(`Send failed.`));
+      }
+
+      const { open } = await prompts([
+        {
+          type: "confirm",
+          name: "open",
+          message: `Open in explorer?`,
+          initial: true,
+        },
+      ]);
+
+      if (open) {
+        const xrplNetwork = network === "xrpl:livenet" ? "xrpl" : "testnet";
+        const url = `https://${xrplNetwork}.xrpl.org/transactions/${status.result.hash}/detailed`;
+        sysOpen(url);
+      }
+
+      await context.coins.disconnect();
+
+      break;
+    }
     case "receive": {
       await gatekeep(context);