+++ identity-wallet/setler-cli/src/kudos.js
@@ -2,6 +2,6 @@
 //  id: string
 //  weight: number 0-1 (optional), will be scaled according to application, but 1 = full, 0 = none
 export const creatorPool = [
-  { id: "did:kudos:github:loremlabs", weight: 1 },
+  { id: "did:kudos:email:setler@loremlabs.com", weight: 1 },
   //    {id: 'did:kudos:email:mankins+3@gmail.com', weight: .1},
 ];